Move old testing for new, change #rebalance
This commit moves the former main.rb testing to a main-old.rb for
reference, as a given driver script is required for the project and will
be the new contents of main.rb.

Additionally, #rebalance did not update @array for Tree instances, and
this commit ensures it updates when a rebalance occurs.